[Refactor:#84] - 콘솔제거 및 폴더 이름 변경
